+++ lib/galaxy/tools/source_store/composite.py
