+++ lib/galaxy/tools/source_store/discover.py
@@ -3,7 +3,7 @@ Tool discovery utilities.
 
 Walks Galaxy's tool configuration files to enumerate every ``<tool>`` referenced
 from any tool_conf without booting a full ``ToolBox``. Used by the populator
-(``galaxy.tool_source_store.populator``) and by callers that need to compare
+(``galaxy.tools.source_store.populator``) and by callers that need to compare
 on-disk confs against the indexed tool set (cold-start auto-populate,
 ``reset_shed_tools``).
 """
