+++ lib/galaxy/tools/source_store/factory.py
