--- packages/app/src/galaxy/tool_source_store
@@ -1,1 +1,0 @@
-../../../../lib/galaxy/tool_source_store
