+++ test/unit/app/tools/source_store/test_composite_store.py
@@ -5,13 +5,13 @@ import tempfile
 
 import pytest
 
-from galaxy.tool_source_store import StoredToolSource
-from galaxy.tool_source_store.composite import CompositeToolSourceStore
-from galaxy.tool_source_store.index import (
+from galaxy.tools.source_store import StoredToolSource
+from galaxy.tools.source_store.composite import CompositeToolSourceStore
+from galaxy.tools.source_store.index import (
     ToolIndex,
     ToolIndexEntry,
 )
-from galaxy.tool_source_store.sqlalchemy import SqlAlchemyToolSourceStore as SqliteToolSourceStore
+from galaxy.tools.source_store.sqlalchemy import SqlAlchemyToolSourceStore as SqliteToolSourceStore
 
 
 @pytest.fixture
