+++ test/unit/scripts/tool_source/test_whoosh_dir.py
@@ -4,7 +4,7 @@ from pathlib import Path
 galaxy_root = Path(__file__).parent.parent.parent.parent.parent
 sys.path.insert(0, str(galaxy_root / "lib"))
 
-from galaxy.tool_source_store.populator import (
+from galaxy.tools.source_store.populator import (
     DEFAULT_STORE_NAME,
     whoosh_dir_for_store,
 )
